Allow revert resize from ERROR
Fixes D-08031

If an instance fails to resize, the instance is moved to error with
neither the old nor new instances running. The old instance often
can be restarted, but the OpenStack Compute API complicates that.
As a workaround, allow revert resize from ERROR if the task_state is
set correctly to indicate it failed during a resize.

NOTE(matiu): When doing the rebase in Feb 9th 2013, I reordered the decorators
in nova/compute/api.py to make sure that wrap_instance_fault is always next
to the function, as it extracts 'instance' from the kwargs and that sometimes
doesn't work if it's not right next to the func. It also makes tests fail.

NOTE(comstud): Squashed this into here:

Leave vm_state and task_state checking to superclass revert_resize

This avoids duplicating logic that needs to be changed twice