Fix collision between conductor and local bw patch.
Change 1df9f6a2d717e9249aab9c61448d2b11442d4803 added a patch
to remove bw data from the 'immediate' .exists events in our
development branch (for billing). When the conductor code landed
in trunk, it changed the way notify_usage_exists was being called
(now it's called through the conductor api) and this was breaking
the instance_usage_audit periodic task that eimits end-of-day
usage events, because there was an extra argument the conductor
code wasn't expecting.

This fixes the conductor api to work with our local patch.