Fix info_cache updates
This fixes info_cache updates to correctly not bomb the API cell.  Every
time we get networking info, the cache is updated to make sure it is in sync.
This would cause a message to the API cell.

Because there's a periodic task on computes that heals the info cache by
getting networking info on all instances... this would bomb the API.

The API cell gets healed differently.. by the cells periodic instance healing
code.

So this patch stops updating the API cell every time we retrieve network info
for an instance... but allows normal updates of the cache to work.